refactor(#255): extract security.rs (scrub_db_args, sanitize_for_display) Move SENSITIVE_FLAGS/ATTACHED_PREFIXES/CONFIG_FILE_FLAGS constants, TokenAction enum, classify_token(), scrub_db_args(), and sanitize_for_display() from mod.rs into a new crates/rskim/src/cmd/security.rs module. Re-export scrub_db_args and sanitize_for_display from mod.rs so all callers (including cmd/log.rs via crate::cmd::sanitize_for_display) continue to work without changes outside cmd/. Moves classify_token and scrub_db_args/sanitize tests to security.rs::tests. format_analytics_label tests remain in mod.rs (that function stays in mod.rs). Co-Authored-By: Claude <noreply@anthropic.com>
